change cost consts to float

+++ mcts/mcts_parameters.h
@@ -49,7 +49,7 @@ struct MctsParameters{
       double ACTION_FILTER_FACTOR;
       std::vector<bool> USE_COST_THRESHOLDING;
       std::vector<bool> USE_CHANCE_CONSTRAINED_UPDATES;
-      std::vector<double> COST_THRESHOLDS;
+      std::vector<float> COST_THRESHOLDS;
       bool USE_LAMBDA_POLICY;
   };
